requirements.txt
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.2
grandalf==0.8
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
azure-functions

